+++ classes/base_decorator.rb
@@ -1,0 +1,9 @@
+require_relative 'nameable'
+
+class BaseDecorator < Nameable
+  @nameable = Nameable.new
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
