--- classes/basedecorator.rb
@@ -1,9 +1,0 @@
-require_relative 'nameable'
-
-class BaseDecorator < Nameable
-    @@nameable = Nameable.new
-    
-    def correct_name
-        @@nameable.correct_name
-    end
-end
