+++ classes/capitalize_decorator.rb
@@ -1,0 +1,12 @@
+require_relative 'base_decorator'
+
+class CapitalizeDecorator < BaseDecorator
+  def initialize(person)
+    super
+    @person = person
+  end
+
+  def correct_name
+    @person.correct_name.upcase
+  end
+end
