+++ classes/nameable.rb
@@ -1,5 +1,5 @@
 class Nameable
-    def correct_name
-        raise NotImplementedError, 'correct_name method can not be used'
-    end
+  def correct_name
+    raise NotImplementedError, 'correct_name method can not be used'
+  end
 end
