+++ classes/person.rb
@@ -1,7 +1,8 @@
 require_relative 'nameable'
 # suberclass
 class Person < Nameable
-  def initialize(id, age, name = 'unkown', parent_permission: true)
+  def initialize(_id, age, name = 'unkown', parent_permission: true)
+    super
     @nameable = Nameable.new
     @id = Random.rand(1..1000)
     @name = name
@@ -21,6 +22,3 @@ class Person < Nameable
     of_age? || parent_permission
   end
 end
-
-p = Person.new(1,45,'asd')
-print p.correct_name
